# [SPARK-6326][SQL] Improve castStruct to be faster