Current `castStruct` should be very slow. This pr slightly improves it.