@viirya 2015-03-15
Simple benchmark conducting 1000000 the `struct casting` in `ExpressionEvaluationSuite`:

before pr: 59.149s
after pr: 53.869s

Conducting 1000000 the `complex casting` in `ExpressionEvaluationSuite`:

before pr: 33.975s
after pr: 31.879s

It is slightly improved.