@viirya 2015-03-15
With new commit.

Conducting 1000000 the `struct casting` in `ExpressionEvaluationSuite`:

before pr: 59.149s
after pr: 47.243s

Conducting 1000000 the `complex casting` in `ExpressionEvaluationSuite`:

before pr: 33.975s
after pr: 27.51s