@marmbrus 2015-03-26
Merged to master. Thanks!